question on the second paragraph
You say you used keywords like multiscale modeling and attended the conference, but that the people there were not the ones you would normally encounter. This seems confusing -- if the keyword was something you were interested in, I'd expect you to have known about the conference and the journals the people published in.