Fix definition of character and competence in text
The second to last paragraph had an incorrect mention of the internal and external relationship of character and competence.  While character is certainly about who a person really is, the measure of character is an external reference to those best properties that benefit society.  Similarly, competence is a measure of skill, independent of its intrinsic value to society.