Some additions and corrections
Made m3 in the comparison table into a proper superscript.
Added placeholder for Author Bio.
Added References heading.
Added BSSw metadata, set for preview.

Note that the actual references are not showing up in the GitHub preview of the document.  We'll need to make sure they appear on the BSSw site.

I'm guessing Matt may take issue with the extensive referencing.  I agree that its not how most of our blog articles look, but I think it is appropriate here.  And besides, Mark has done them in a way that is pretty unobtrusive.  If you don't scroll down to the bottom of the article, you might not even notice they're there.