Extend content for 'Working Effectively with Legacy Code' (#331)
I feel that most people who look at this article will not read the book.
Therefore, I want to give them the crtical ideas in the book to at least start
to educate them about this critical topic.  This is not that long and a person
can read in 5 minutes and it is easy to browse.