Make event template consistent with current site needs
Maybe I've overdone the explanations, but I think this kind of information is going to be necessary if we want to get inputs that are publishable with a minimum of effort.  Another way to approach this might be a more detailed contributor's guide for each type of content that basically explains the information needed and how to expect it to be presented on the site, along with appropriate style guidance.