Corrected dashes in deck
It seems like somehow a couple of emdashes got introduced here in the original text rather than the hyphens we'd normally expect. Then Gail added hyphens with the intent of writing '--', as is used elsewhere and they look strange because they don't line up.  Since I don't know how to get the emdashes everywhere, I changed these to double hyphens to be consistent with the rest of the article.