Styling
Moved the description of the article to the deck resource text.
Italicized article name in deck resource text. 
Added resource table; need information on how the focus should be classified.
Created a section for main body text; need information on the resource. 
Put the contribution details at the bottom of the article.