Grammar/Punctuation Edits
Edits regarding grammar/punctuation, etc.

Also, changed the syntax of a good portion of this paragraph 
"Several reasons have given way to this scenario. Initially, the models that dominated enterprise software before the advent of Agile methods had far too many process overheads to be even remotely feasible in science environments. More importantly, two very important ideologies coincided with one another: the growing popularity of Agile principles in enterprise software and the scientific software community's growing awareness of the adverse impact on productivity in the absence of defined software processes.  However, the ways in which the adoption of Agile principles has affected these two different areas of software development could not be more different."

I found the original wording to be somewhat verbose and difficult to understand. Please check to verify that the content is being presented properly, or feel free to revert back to the original.