Initial draft of blog post. (#701)
* Initial draft of blog post.
This will need to have the img src url fixed after the PR in the image repo is
merged.

* Fix 2nd unordered list

* Edits to Useful Practices

Standard edits for clarity, fixed spacing issues, checked all links. 

Should the IEEE glossary be a link? 
10.1109/IEEESTD.1990.101064 or https://ieeexplore.ieee.org/document/159342

* Update 2021-03-useful-practices-for-SEoMsDSP.md

I tried to add catchy subheads to attract attention and break up the paragraphing.
I also suggest limiting what seems to me to be extra information (like names of places not really needed to appreciate the point) and acronyms that are not used again.

* Accepting most of Gail's changes + minor changes

* Update metadata

* Metadata, hero image

* Debug hero image

* Reconverting the supposed typo :)

* Continue with debugging of hero image

* Update 2021-03-useful-practices-for-SEoMsDSP.md

This re-wording of the end of the "Challenges facing a medium-sized project" is entirely optional.

* Update 2021-03-useful-practices-for-SEoMsDSP.md

Co-authored-by: Kasia Salim <65253296+karbarz@users.noreply.github.com>
Co-authored-by: gpieper <pieper@mcs.anl.gov>
Co-authored-by: Rinku Gupta <rgupta@mcs.anl.gov>