Only merge PR branch to 'preview' if contains 'preview' label (#744, #330)
To get this to work, I had to add the if statement looking for the 'preview'
label.  The {label_name: 'preview'} setting for the devmasx/merge-branch
action did not work (it just merged the PR branch to 'preview' even if the PR
did not have the 'preview' label).  I also had to add 'labeled' to the
pull_request_target types triggers so that adding the 'preview' label to a PR
that does not change in any other way will trigger the sync of the PR to
'preview'.

I also changed the clone of the repo to get the 'preview' branch instead of
the default 'master' branch.  Not that big of a deal but it seems confusing to
checkout the 'master' branch when you are really just wanting to merge to the
'preview' branch.

I added a comment at the bottom for the non-obvious parts of the file.