Template for bssw.io PRs (#694, #702, #744, #750, #330)
This adds a single template file for PRs.  It uses a meta-template for three
different types of PRs with three different checklists with instructions on
how to quickly modify the template for the type of PR.  (Instructions should
be straightforward).

I had tried using different template *.md files for each PR, but GitHub
requires you to **manually** edit the URL for the creation of the PR to add
`&template=<template-file-name>` to the end of the URL.  That is just way to
complicated and error-prone.  Therefore, I choose the to provide a single
meta-template *.md file which will be used by default that can be quickly
edited for the type of PR being created.  Hopefully this is the lesser of
multiple evils (i.e. no checklist template at all or hard-to-use template
files you have to manually add to the PR-creation URL).