Update docs for updated citations/references policy (#702) (#750)
* Fix spelling of 'Michael'

* Run wikize_refs.py and small manual fixups

Now wikize_refs.py should run without modification on this file.  And I can
also get rid of the *-src.md file.

* Remove *-src.md now that wikize_refs.py can be run in-place

* Remove off-page hyperlinks from main text (#702)

This uses the new guidelines for clean citations in a markdown file.

* Update auto-generated reference stuff with wikize_refs.py (#703)

* Fix some misspellings (#702)

I saw these as I was getting ready to add section on citations (see #702).

Also, the accepted capitalization is "GitHub", not "Github".

* Add install of gem kramdown-parser-gf (#702)

I needed to do this in order to get Jekyll to work locally for me.

* Ignore locally generated jekyll files (#702)

When I run jeykll locally, it generates these files and we don't want to
version control them.

* Update for updated behavior of wikize_refs.py for in-place running (#718)

The document was describing the old behavior before PR #718.

* Improve short wikize_refs.py doc a little in README.md file (#702)

Just enough to do what you need and nothing more.  But this links to more
expansive documentation.

* Add citations/references policy and update content publishing (#702)

* Address for conflict from commit 9b0eb0af not merged to 'master' (#702)

It appears the commit 9ae77293 was never merged to 'master' but was merged
into 'preview' using the merge commit 92145cad.  Not sure what happened there.

* Add pointer to example for formal citations/references (#702)

This responds to feedback from @markcmiller86 to PR #750.

NOTE: This points to the 'master' branch for the example file
WhenNotToWriteAutomatedTests.md.  This file is updated in PR #750 so this link
will only point to the updated *.md file until PR #750 gets merged.

* Change wording 'Ensure to run X' to 'Ensure X is run' (#702)

Responding to feedback from @markcmiller86 in PR #750.

Co-authored-by: Rinku Gupta <rgupta@mcs.anl.gov>