Add new original short article for git beginners tips (#560)
This was copied from:

* https://ideas-productivity.org/resources/howtos/git-tutorial-and-reference-collection/beginner-tips/

by first downloading PDF and then using an online converter
https://pdf2md.morethan.io/.  I had to hand modify the converted text quite a
bit (because mistakes were made in the translation) and I had to manually put
back in the hyperlinks.  But I had to read over the entire document again
anyway so that was fine.

I verified that all of the links still work.