The 'S' in Git LFS stands for 'Storage' (#560)
This text make it seem that the 'S' in Git LFS stood for 'Support' but it
actually stands for 'Storage'.

Also, technically, git-lfs is not officially supported by the core Git
development team that puts out Git releases.  Instead, it is collaboration
between mostly GitHub, GitLab, and (Atlassian) BitBucket.  So it is not
"Git's" tool.  Instead, it is a tool built on top of Git and integrated into
Git.  That is why I removed the "'s" from "Git's Large File ...".