Add explicit HTML anchors for section names (#560)
Seems the custom Markdown translator for the bssw.io site does not maintain
the section header links like GitHub markdown does.  This is an attempt to
manually put in these anchors.

If this works, we may re-think how to do this type of thing (and may not use
short section headers like this).