Add beginning of section on non-standard handling of Markdown by bssw.io
This was motivated by the few hours of time we wasted by assuming that
subsection links worked as part of PR #823.

Hopefully a reminder like this will help avoid wasted time in the future.
(And if people get caught by this again, then we can say it was their fault
for not reading the documentation :-)).

This is just the beginning of the list of issues.  People with more knowledge
that us should add to this list before the PR gets merged to 'master'.