Switch to subsubsection headers and shorten header names
This has the advantages that these items have their own anchors/links and they
are shown in the table of contents.

But it has the disadvantage that these are shown in the table fo contents
(which seems to make this section out-sized compared to the other sections on
the page).

Therefore, we may not want to do this.  I will let other people decide.