Fix link to 'utils/README.md#wikize_refspy' from Jekyll-generated page (#927)
I don't see what other choice you have but to put in an absolute URL so this
works from the Jekyll-generated page.  NOTE: This link will also work from the
`*.md` file as well it just will still be absolute and not point to the file
on the same branch.  Hopefully that is not a big deal.