Update documentation for move to images/ subdir (#703)
NOTE: I avoided updating documentation for the handling of images with the
"slides" in the Homepage.md file in bssw_content_highlighting.md.  The work to
update the handling of this file is in-progress and there is no sense in
updating this until they are done.  Plus, bssw_content_highlighting.md is not
linked to from the main menu so it is doubtful people people would stumble on
it by accident so it is not a priority to update.  (I only found it by running
an `find ... grep` operation.)