Inplace wikize the ref (#523)
Note, I produced this with:

  wikize_refs.py -r -1 -i TheDeveloperCertificateOfOrigin.md

For some reason, it did not renumber if I ran with '-r 1'.  But If I run
wikize_ref.py again, it gives the error:

  Some footnotes never appear in the references...
  ['11', '1', '10', '4', '8', '9', '14', '2', '13', '7', '5', '0', '3', '6', '12']
  Either correct above errors or run with -w (--warn)

That is nonsense because those footnotes do and references do exist just fine
from what I can tell.  So I can't run the tool again.