Remove < ... > that is dropped by bssw.io generator (#523)
The bssw.io site generator incorrectly slips the text <any-author> inside of:

  `git commit --amend --author "<any-author>"`

The bssw.io site generator seems to be incorrectly treating <any-author> as an
HTML element but it should not because it is inside of `...`.  The GitHub
Markdown viewer does not do that.  (Just another unique quirk of the bssw.io
site generator.)