Update 2022-05-build-integration-testing.md
Does Anshu in  her bio  wish to mention Flash-X?
I left  the  titles in the bios in caps but usually they are not, especially when  one says "a Principal.." or "a  Computer..."

I would  prefer for the  blog something  more  exciting  then the heading "Introduction" -- perhaps  Software Challenges in a Changing Landscape. Or you could  omit any heading.

I  also note that  in the first  few responses  to the question (about software the panelists  have  been involved in), the panelists state their role. The  next few, however, just state  the software. I'd expect at least  a brief note about their role (lead?  tester?) 

In Ross's first respons,  is the package  he mentions  xSDK? If  so , perhaps  use  that  rather than spelling  out  the  entire project name -- since it was spelled out in  the preceding intro.