Remove double [[.]] around references, add titles to all refereces
Turns out the wikize-refs.py tool cannot handle adding an extra bracket [.]
around references [[.]].  I had to remove all of those (even in places where
it would have been nice to keep them).

I also added titles to all references when I was trying to figure out what was
wrong with wikize-refs.py but it was unrelated to the problem.  (But I decided
to keep them.)