Switch to H3 and bold for headings
Based on suggestion by @bernhold, I have made the following changes:

* H2 ## => H3 ####
* H3 ### => **bold**

This is to try to fix the sizes of the fonts for the subsection and
subsubsections.  Currently, the H2 ## heading text looks as large as the H1 #
text used for the title when this gets published on bssw.io.