Some revisions
I think the original and proposed text was possibly conflating a few different things like actual content vs. ideas for content. It introduced concept of abstract and reviews, etc and seemed to describe a heavier handed "ideas gathering" process than I think we are really after. It was also a tad repetitive in description of process here. So, am proposing this revision