Remove target 'master' branch constraint
There are cases where we want allow spacked PR and be able to 
view the overall result on the preview.bssw.io site.  To allow for that
we need to be able to merge branches to 'preview' even when the target is not 'master'.

For example, this is stopping the preview of Pr #1679.