Further small edits
Technically a code review does not fix any defects that are found.
Fixing the defects is part of the follow-up process.  Code reviewers
generally should not be fixing problems they find themselves.

But I did leave in that this process leads to fixing the defects found.

Also, I just needed to add a commit to test the updated
merge-to-preview.yml workflow.