Change from 'master' to 'main' (#674)
This should update all of the workflows, documetation, and other documents
that depend on the 'master' branch in the bssw.io repo.  There are other hits
for 'master' that are not related to the bssw.io 'master' branch that were not
updated (a manual process unfortunately).

FYI: If I were to do this refactoring again, I would break it up into two
commits.

The first commit would replace:

  github.com/betterscientificsoftware/bssw.io/blob/master/
  github.com/betterscientificsoftware/bssw.io/tree/master/
  github.com/betterscientificsoftware/bssw.io/raw/master/

with:

  github.com/betterscientificsoftware/bssw.io/blob/main/
  github.com/betterscientificsoftware/bssw.io/tree/main/
  github.com/betterscientificsoftware/bssw.io/raw/main/

using a script.

The second commit would do the rest of the manual replacements of 'master'
with 'main'.