Basic copy and formatting of text
This is a straight copy of text from the provided MS Word document into the
standard bssw.io Blog article template.  Not sure how subsubsections will look
on preview.bssw.io but I will look at that when we get to that point.

There is still some other info that needs to be filled in (see "???").

Signed-off-by: Roscoe A. Bartlett <rabartl@sandia.gov>