Minor edits for publication
For better or worse, the date in the metadata needs to be an exact date to define the transition from upcoming event to past event. So I've chosen Friday 19 December.  We can easily change this if desired.  And you can update it (with another PR) if you want to extend the deadline later on.